+++ src/main/java/gameobj/staticobj/Wall.java
@@ -8,7 +8,7 @@ public class Wall extends StaticObj {
     /**
      * Constructor for a wall; utilises StaticObj's constructor.
      */
-    public Wall(int posX, int posY, int width, int height, PImage sprite) {
+    public Wall(double posX, double posY, int width, int height, PImage sprite) {
         super(posX, posY, width, height, sprite);
     }
 }
